Patch gates for beeg knight
