+++ SmolKnight.csproj
@@ -99,7 +99,8 @@
         <Compile Include="Properties\AssemblyInfo.cs" />
 
         <Compile Include="SmolKnight\Menu\ModMenu.cs" />
-        
+
+        <Compile Include="SmolKnight\Patch\GatePatcher.cs" />        
         <Compile Include="SmolKnight\Patch\ShinyItemStandPatcher.cs" />
         <Compile Include="SmolKnight\Patch\VignettePatcher.cs" />
         <Compile Include="SmolKnight\Patch\PlayerDataPatcher.cs" />
